Avoid the warning about a double input to abs()
The patch keeps the current behavior, but the warning was much
probably actually raising a bug or a small conceptual mistake:
the type of L_xx.size is float (see tab.h), however it is always
filled with int values because it is initialized to 0 and updated
unsing "int abs(int);". Worse, the input of abs() is always a
double! There's obviously some inconsistency there...

+++ source/traiter.c
@@ -580,7 +580,7 @@ static int *tab_sort_rows_xx(
 	d = piplib_int_get_d(Denom(tp, i));
 	for (j = 0; j < nvar; j++) {
 	    t = piplib_int_get_d(Index(tp,i,j))/d;
-	    s = piplib_max(s, abs(t));
+	    s = piplib_max(s, abs((int)t)); // NdCed: or fabs(t) ?
 	}
 	tp->row[i].size = s;
 	smax = piplib_max(s, smax);
